requirements.txt
tox
cookiecutter
zsl
Jinja2
PyYAML
future
